flight-service/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==38.0.3
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==23.2
django-nose==1.4.7
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.9.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
nose==1.3.7
oauthlib==3.2.2
openapi-codec==1.3.2
Pillow==9.3.0
platformdirs==3.0.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyotp==2.9.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.19.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stripe==6.2.0
typing_extensions==4.7.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.19.0
whitenoise==6.4.0
